Closures

+++ src/native.rs
@@ -1,8 +1,8 @@
 use std::time::{SystemTime, UNIX_EPOCH};
 
-use super::value::Value;
+use crate::value::Value;
 
-pub fn clock_native(_arg_coun: u8, _values: &[Value]) -> Value {
+pub fn clock_native(_arg_coun: usize, _values: &[Value]) -> Value {
     let time = SystemTime::now()
         .duration_since(UNIX_EPOCH)
         .unwrap()
